Experiment: use tick view for visuals, u8 -> f32.

+++ examples/no-window/listener.rs
@@ -3,7 +3,7 @@ use saunter::listener::Listener;
 use crate::tick::NoWindowTick;
 
 pub struct NoWindowListener {
-    pub val: f32,
+    pub val: u8,
 }
 impl Listener for NoWindowListener {
     type TickType = NoWindowTick;
@@ -16,7 +16,7 @@ impl Listener for NoWindowListener {
         _events: &mut Vec<saunter::event::Event<Self::EventType>>,
         time: std::time::Instant,
     ) -> Result<Self::TickType, saunter::error::SaunterError> {
-        self.val = 1.0 - self.val;
+        self.val = 1 - self.val;
         log::info!("{}", self.val);
 
         Ok(NoWindowTick { val: self.val, time })
